+++ daemon/src/collab_settlement_maker.rs
@@ -1,4 +1,3 @@
-use crate::cfd_actors::load_cfd;
 use crate::command;
 use crate::maker_inc_connections;
 use crate::process_manager;
@@ -56,7 +55,11 @@ impl Actor {
             );
 
             let mut conn = self.db.acquire().await?;
-            let cfd = load_cfd(self.proposal.order_id, &mut conn).await?;
+
+            // Collaborative settlement does not fit into the command abstraction which I think
+            // roots in its asymmetric design.
+            #[allow(deprecated)]
+            let cfd = crate::command::load_cfd(self.proposal.order_id, &mut conn).await?;
 
             let settlement =
                 cfd.sign_collaborative_settlement_maker(self.proposal, msg.sig_taker)?;
